# [WIP] [tabular] speed up tests