--- a/tabular/src/autogluon/tabular/models/lgb/lgb_model.py
+++ b/tabular/src/autogluon/tabular/models/lgb/lgb_model.py
@@ -528,6 +528,10 @@ class LGBModel(AbstractModel):
         num_gpus = 0
         return num_cpus, num_gpus
 
+    @classmethod
+    def supported_problem_types(cls) -> list[str] | None:
+        return ["binary", "multiclass", "regression", "quantile", "softclass"]
+
     @property
     def _features(self):
         return self._features_internal_list