--- a/tabular/tests/unittests/models/test_knn.py
+++ b/tabular/tests/unittests/models/test_knn.py
@@ -1,25 +1,8 @@
 from autogluon.tabular.models.knn.knn_model import KNNModel
 
 
-def test_knn_binary(fit_helper):
-    fit_args = dict(
-        hyperparameters={KNNModel: {}},
-    )
-    dataset_name = "adult"
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+def test_knn(fit_helper):
+    model_cls = KNNModel
+    model_hyperparameters = {"n_neighbors": 2}
 
-
-def test_knn_multiclass(fit_helper):
-    fit_args = dict(
-        hyperparameters={KNNModel: {}},
-    )
-    dataset_name = "covertype_small"
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
-
-
-def test_knn_regression(fit_helper):
-    fit_args = dict(
-        hyperparameters={KNNModel: {}},
-    )
-    dataset_name = "ames"
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    fit_helper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters, bag="first", refit_full="first")