--- a/tabular/tests/unittests/models/test_knn.py
+++ b/tabular/tests/unittests/models/test_knn.py
@@ -5,4 +5,4 @@ def test_knn(fit_helper):
     model_cls = KNNModel
     model_hyperparameters = {"n_neighbors": 2}
 
-    fit_helper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters, bag="first", refit_full="first")
+    fit_helper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters)