*Issue #, if available:*

*Description of changes:*

WIP: Do not merge.

Massively speed up unit tests for tabular models

- Switch from real datasets to toy synthetic datasets
- Use toy model hyperparameters when faster
- Avoid calling `predictor.info()` for each model fit
- Avoid calling `predictor.leaderboard(..., extra_info=True)` for each model fit
- Only call refit_full for a given model family once, no need to test refit_full on every dataset
- Only test bagging for a given model family once, no need to test bagging on every dataset
- Ensure we test bagging for every model family (previously was inconsistent)
- Ensure we are testing every supported problem_type (previously it was inconsistent)

## Speedup

- LightGBM tests: 130s -> 16s 
- Dummy tests: 23s -> 12s
- NeuralNetTorch tests: 140s -> 8s
- NeuralNetFastAI tests: 15s -> 2s
- RandomForest tests: 87s -> 8s
- ExtraTrees tests: ?s -> 7s
- XGBoost tests: ?s -> 4s
- CatBoost tests: ?s -> 3s
- Linear tests: ?s -> 6s

TODO:

- [ ] Speed up remaining models
- [ ] Speed up `test_tabular.py`
- [ ] Adopt a generic model test function?
- [ ] Verify categorical, missing value, etc. handling for models?

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.